// code/backend/koajs/koajs_search_article/src/generated-clients/searcharticleservice/models/SearchArticleRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Solr Search Article Service
 * Solr Search Article Service
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {exists} from '../runtime';
import type {SearchCustomer} from './SearchCustomer';
import {SearchCustomerFromJSON, SearchCustomerToJSON,} from './SearchCustomer';

/**
 *
 * @export
 * @interface SearchArticleRequest
 */
export interface SearchArticleRequest {
    /**
     *
     * @type {string}
     * @memberof SearchArticleRequest
     */
    q?: string;
    /**
     *
     * @type {number}
     * @memberof SearchArticleRequest
     */
    offset?: number;
    /**
     *
     * @type {number}
     * @memberof SearchArticleRequest
     */
    limit?: number;
    /**
     *
     * @type {SearchCustomer}
     * @memberof SearchArticleRequest
     */
    customer?: SearchCustomer;
}

/**
 * Check if a given object implements the SearchArticleRequest interface.
 */
export function instanceOfSearchArticleRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function SearchArticleRequestFromJSON(json: any): SearchArticleRequest {
    return SearchArticleRequestFromJSONTyped(json, false);
}

export function SearchArticleRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SearchArticleRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {

        'q': !exists(json, 'q') ? undefined : json['q'],
        'offset': !exists(json, 'offset') ? undefined : json['offset'],
        'limit': !exists(json, 'limit') ? undefined : json['limit'],
        'customer': !exists(json, 'customer') ? undefined : SearchCustomerFromJSON(json['customer']),
    };
}

export function SearchArticleRequestToJSON(value?: SearchArticleRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {

        'q': value.q,
        'offset': value.offset,
        'limit': value.limit,
        'customer': SearchCustomerToJSON(value.customer),
    };
}