// code/backend/koajs/koajs_search_article/src/generated-clients/customerpriceservice/models/CustomerPriceEntry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CustomerPrice Service
 * CustomerPrice Service
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {exists} from '../runtime';

/**
 *
 * @export
 * @interface CustomerPriceEntry
 */
export interface CustomerPriceEntry {
    /**
     *
     * @type {number}
     * @memberof CustomerPriceEntry
     */
    id?: number;
    /**
     *
     * @type {number}
     * @memberof CustomerPriceEntry
     */
    customerId?: number;
    /**
     *
     * @type {number}
     * @memberof CustomerPriceEntry
     */
    discount?: number;
    /**
     *
     * @type {number}
     * @memberof CustomerPriceEntry
     */
    startYear?: number;
    /**
     *
     * @type {number}
     * @memberof CustomerPriceEntry
     */
    endYear?: number;
    /**
     *
     * @type {Date}
     * @memberof CustomerPriceEntry
     */
    created?: Date;
}

/**
 * Check if a given object implements the CustomerPriceEntry interface.
 */
export function instanceOfCustomerPriceEntry(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function CustomerPriceEntryFromJSON(json: any): CustomerPriceEntry {
    return CustomerPriceEntryFromJSONTyped(json, false);
}

export function CustomerPriceEntryFromJSONTyped(json: any, ignoreDiscriminator: boolean): CustomerPriceEntry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {

        'id': !exists(json, 'id') ? undefined : json['id'],
        'customerId': !exists(json, 'customerId') ? undefined : json['customerId'],
        'discount': !exists(json, 'discount') ? undefined : json['discount'],
        'startYear': !exists(json, 'startYear') ? undefined : json['startYear'],
        'endYear': !exists(json, 'endYear') ? undefined : json['endYear'],
        'created': !exists(json, 'created') ? undefined : (new Date(json['created'])),
    };
}

export function CustomerPriceEntryToJSON(value?: CustomerPriceEntry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {

        'id': value.id,
        'customerId': value.customerId,
        'discount': value.discount,
        'startYear': value.startYear,
        'endYear': value.endYear,
        'created': value.created === undefined ? undefined : (value.created.toISOString()),
    };
}